feat: special handling for `map[string]any`
Since `map[string]any` is commonly used for cases such as JSON handling, this change handles it in a way that prevents the heap allocations normally associated with map iteration using reflect.
This leads to the value being forced to the heap, adding an additional heap allocation to all cases even if no map is contained.